+++ includes/class-newspack-listings-blocks.php
@@ -10,6 +10,8 @@
 namespace Newspack_Listings;
 
 use \Newspack_Listings\Newspack_Listings_Core as Core;
+use \Newspack_Listings\Newspack_Listings_Products as Products;
+use \Newspack_Listings\Newspack_Listings_Settings as Settings;
 use \Newspack_Listings\Newspack_Listings_Taxonomies as Taxonomies;
 
 defined( 'ABSPATH' ) || exit;
@@ -87,26 +89,37 @@ final class Newspack_Listings_Blocks {
 			];
 		}
 
+		$localized_data = [
+			'post_type_label'    => $post_type_label,
+			'post_type'          => $post_type,
+			'post_type_slug'     => array_search( $post_type, Core::NEWSPACK_LISTINGS_POST_TYPES ),
+			'post_types'         => $post_types,
+			'taxonomies'         => $taxonomies,
+			'currency'           => function_exists( 'get_woocommerce_currency' ) ? get_woocommerce_currency() : __( 'USD', 'newspack-listings' ),
+			'currencies'         => function_exists( 'get_woocommerce_currencies' ) ? get_woocommerce_currencies() : [ 'USD' => __( 'United States (US) dollar', 'newspack-listings' ) ],
+
+			// If we don't have ANY listings that can be added to a list yet, alert the editor so we can show messaging.
+			'no_listings'        => 0 === $total_count,
+			'date_format'        => get_option( 'date_format' ),
+			'time_format'        => get_option( 'time_format' ),
+
+			// Self-serve listings features are gated behind an environment variable.
+			'self_serve_enabled' => defined( 'NEWSPACK_LISTINGS_SELF_SERVE_ENABLED' ) && NEWSPACK_LISTINGS_SELF_SERVE_ENABLED,
+		];
+
+		if ( $localized_data['self_serve_enabled'] ) {
+			$localized_data['self_serve_listing_types']      = Products::get_listing_types();
+			$localized_data['self_serve_listing_expiration'] = Settings::get_settings( 'newspack_listings_single_purchase_expiration' );
+
+			if ( Products::is_listing_customer() ) {
+				$localized_data['is_listing_customer'] = true;
+			}
+		}
+
 		wp_localize_script(
 			'newspack-listings-editor',
 			'newspack_listings_data',
-			[
-				'post_type_label'    => $post_type_label,
-				'post_type'          => $post_type,
-				'post_type_slug'     => array_search( $post_type, Core::NEWSPACK_LISTINGS_POST_TYPES ),
-				'post_types'         => $post_types,
-				'taxonomies'         => $taxonomies,
-				'currency'           => function_exists( 'get_woocommerce_currency' ) ? get_woocommerce_currency() : __( 'USD', 'newspack-listings' ),
-				'currencies'         => function_exists( 'get_woocommerce_currencies' ) ? get_woocommerce_currencies() : [ 'USD' => __( 'United States (US) dollar', 'newspack-listings' ) ],
-
-				// If we don't have ANY listings that can be added to a list yet, alert the editor so we can show messaging.
-				'no_listings'        => 0 === $total_count,
-				'date_format'        => get_option( 'date_format' ),
-				'time_format'        => get_option( 'time_format' ),
-
-				// Self-serve listings features are gated behind an environment variable.
-				'self_serve_enabled' => defined( 'NEWSPACK_LISTINGS_SELF_SERVE_ENABLED' ) && NEWSPACK_LISTINGS_SELF_SERVE_ENABLED,
-			]
+			$localized_data
 		);
 
 		wp_register_style(